+++ perl/FSM/Pipeline/HDLGenerator.pm
@@ -38,6 +38,7 @@ use FSM::IR::StructuralRTLIR::ConnectionExpr qw(
     normalized_binding
     binding_expr
     binding_signal_name
+    binding_signal_names
     binding_expr_text
 );
 use FSM::SourceClassifier;
@@ -2451,11 +2452,13 @@ sub composition_system_signal_names ($self, $composition_plan) {
     if ((!defined($clock_name) || !length($clock_name)) || (!defined($reset_name) || !length($reset_name))) {
         for my $instance (@{$composition_plan->instances || []}) {
             my %bindings = map {
-                (($_->{port_name} || '') => binding_signal_name($_))
+                (($_->{port_name} || '') => [ @{binding_signal_names($_)} ])
             } @{$instance->port_bindings || []};
 
             for my $port (@{$instance->interface_ports || []}) {
-                my $bound_signal = $bindings{$port->name} || next;
+                my $bound_signals = $bindings{$port->name} || next;
+                next unless ref($bound_signals) eq 'ARRAY' && @$bound_signals == 1;
+                my $bound_signal = $bound_signals->[0] || next;
                 my $type = $port->type || '';
                 $clock_name ||= $bound_signal if $type eq 'clock';
                 $reset_name ||= $bound_signal if $type eq 'reset';
@@ -3442,8 +3445,8 @@ sub build_composition_shared_datapath_candidates ($self, $composition_plan, $str
     for my $instance (@{$structural_rtl_ir_hash->{instances} || []}) {
         next unless (($instance->{kind} || '') eq 'fsmc');
 
-        my %bindings = map {
-            (($_->{port_name} || '') => binding_signal_name($_))
+        my %binding_signal_sets = map {
+            (($_->{port_name} || '') => [ @{binding_signal_names($_)} ])
         } @{$instance->{port_bindings} || []};
         my $child = $child_by_instance{$instance->{instance_name}} || {};
         my %drive_family_by_signal = map {
@@ -3463,13 +3466,16 @@ sub build_composition_shared_datapath_candidates ($self, $composition_plan, $str
             my $output_drive_family = ref($drive_family) eq 'HASH'
                 ? _clone_structured_value($drive_family)
                 : {};
+            my $bound_signals = $binding_signal_sets{$port->{name}} || [];
+            my $bound_signal = @$bound_signals == 1 ? $bound_signals->[0] : '';
             push @{$candidate_groups{$key}{contributors}}, {
                 kind => ($child->{kind} // $instance->{kind}),
                 instance_name => ($child->{instance_name} // $instance->{instance_name}),
                 module_name => ($child->{module_name} // $instance->{module_name}),
                 source_name => ($child->{source_name} // $instance->{source_name}),
                 endpoint => (($instance->{instance_name} // 'unknown').'.'.($port->{name} // 'unknown')),
-                bound_signal => $bindings{$port->{name}},
+                bound_signal => $bound_signal,
+                bound_signals => [@$bound_signals],
                 intent_hir => ($child->{intent_hir} || {}),
                 lowered_rtl_ir => ($child->{lowered_rtl_ir} || {}),
                 structural_rtl_ir => ($child->{structural_rtl_ir} || {}),
@@ -3494,7 +3500,10 @@ sub build_composition_shared_datapath_candidates ($self, $composition_plan, $str
                 instance_name => ($child->{instance_name} // $instance->{instance_name}),
                 module_name => ($child->{module_name} // $instance->{module_name}),
                 endpoint => (($instance->{instance_name} // 'unknown').'.'.($port->{name} // 'unknown')),
-                bound_signal => $bindings{$port->{name}},
+                bound_signal => (@{$binding_signal_sets{$port->{name}} || []} == 1)
+                    ? ($binding_signal_sets{$port->{name}} || [])->[0]
+                    : '',
+                bound_signals => [@{$binding_signal_sets{$port->{name}} || []}],
             };
         }
     }