+++ t/139-composition-shared-datapath-candidate-metadata.t
@@ -98,6 +98,7 @@ FSM
                         source_name => 'left_src',
                         endpoint => 'left.status_bus',
                         bound_signal => 'left_status',
+                        bound_signals => ['left_status'],
                         output_drive_family => {
                             signal_name => 'status_bus',
                             width => 8,
@@ -144,6 +145,7 @@ FSM
                         source_name => 'right_src',
                         endpoint => 'right.status_bus',
                         bound_signal => 'right_status',
+                        bound_signals => ['right_status'],
                         output_drive_family => {
                             signal_name => 'status_bus',
                             width => 8,